# openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "UFC_borci API",
    "version": "1.0"
  },
  "servers": [{ "url": "http://localhost:3000/api/v1" }],
...
  "paths": {
    "/fighters": {
      "get": {
        "summary": "Get UFC fighters",
        "description": "Fetches all fighter objects from database",
        "parameters": [],
        "tags": ["UFC fighters"],
        "responses": {
          "200": {
            "description": "Fetched all fighter objects",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FighterResponse"
                }
              }
            }
          },

          "500": {
            "$ref": "#/components/responses/ERR_CONNECT"
          }
        }
      },
      "post": {
        "tags": ["UFC fighters"],
        "summary": "Add new UFC fighter",
        "description": "Adds new fighter object to database",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Fighter"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Fighter added successfully!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostSuccessResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ERR_ADD"
          },
          "404": {
            "$ref": "#/components/responses/ERR_OPP"
          },
          "500": {
            "$ref": "#/components/responses/ERR_CONNECT"
          }
        }
      },
      "put": {
        "tags": ["UFC fighters"],
        "summary": "Change an UFC fighter",
        "description": "Modifies values of an existing figter object",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FighterWithId"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Fighter data modified successfully!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "204": {
            "description": "Fighter does not exist!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ERR_UPD"
          },
          "404": {
            "$ref": "#/components/responses/ERR_OPP"
          },
          "500": {
            "$ref": "#/components/responses/ERR_CONNECT"
          }
        }
      },
      "head": {
        "tags": ["Unimplemented"],
        "description": "Not implemented",
        "parameters": [],
        "responses": {
          "501": {
            "$ref": "#/components/responses/ERR_NOT_IMPLEMENTED"
          }
        }
      },
      "patch": {
        "tags": ["Unimplemented"],
        "description": "Not implemented",
        "parameters": [],
        "responses": {
          "501": {
            "$ref": "#/components/responses/ERR_NOT_IMPLEMENTED"
          }
        }
      },
      "options": {
        "tags": ["Unimplemented"],
        "description": "Not implemented",
        "parameters": [],
        "responses": {
          "501": {
            "$ref": "#/components/responses/ERR_NOT_IMPLEMENTED"
          }
        }
      }
    },
    "/fighters/{id}": {
      "get": {
        "tags": ["UFC fighters"],
        "summary": "Get UFC fighter",
        "description": "Fetches specific fighter object from database",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the resource",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fetched fighter object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FighterResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/ERR_FIGHTER_NOT_FOUND"
          },
          "422": {
            "$ref": "#/components/responses/ERR_INVALID_ID"
          },
          "500": {
            "$ref": "#/components/responses/ERR_CONNECT"
          }
        }
      },
      "delete": {
        "tags": ["UFC fighters"],
        "summary": "Delete UFC fighter",
        "description": "Deletes fighter object from database",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the resource",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fighter deleted successfully!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ERR_DEL"
          },
          "422": {
            "$ref": "#/components/responses/ERR_INVALID_ID"
          },
          "404": {
            "$ref": "#/components/responses/ERR_FIGHTER_NOT_FOUND"
          },
          "500": {
            "$ref": "#/components/responses/ERR_CONNECT"
          }
        }
      }
    },
    "/fighters/brief": {
      "get": {
        "tags": ["UFC fighters"],
        "summary": "Get brief UFC fighters",
        "description": "Fetches brief fighter objects from database",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Fetched all fighter objects",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FighterBriefResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ERR_CONNECT"
          }
        }
      }
    },
    "/fights": {
      "get": {
        "tags": ["UFC fights"],
        "summary": "Get UFC fights",
        "description": "Fetches fight objects from database",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Fetched all fight objects",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FightResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ERR_CONNECT"
          }
        }
      },
      "post": {
        "tags": ["Unimplemented"],
        "description": "Not implemented",
        "parameters": [],
        "responses": {
          "501": {
            "$ref": "#/components/responses/ERR_NOT_IMPLEMENTED"
          }
        }
      },
      "put": {
        "tags": ["Unimplemented"],
        "description": "Not implemented",
        "parameters": [],
        "responses": {
          "501": {
            "$ref": "#/components/responses/ERR_NOT_IMPLEMENTED"
          }
        }
      },
      "delete": {
        "tags": ["Unimplemented"],
        "description": "Not implemented",
        "parameters": [],
        "responses": {
          "501": {
            "$ref": "#/components/responses/ERR_NOT_IMPLEMENTED"
          }
        }
      },
      "options": {
        "tags": ["Unimplemented"],
        "description": "Not implemented",
        "parameters": [],
        "responses": {
          "501": {
            "$ref": "#/components/responses/ERR_NOT_IMPLEMENTED"
          }
        }
      },
      "head": {
        "tags": ["Unimplemented"],
        "description": "Not implemented",
        "parameters": [],
        "responses": {
          "501": {
            "$ref": "#/components/responses/ERR_NOT_IMPLEMENTED"
          }
        }
      },
      "patch": {
        "tags": ["Unimplemented"],
        "description": "Not implemented",
        "parameters": [],
        "responses": {
          "501": {
            "$ref": "#/components/responses/ERR_NOT_IMPLEMENTED"
          }
        }
      }
    },
    "/fights/{id}": {
      "get": {
        "tags": ["UFC fights"],
        "summary": "Get UFC fight",
        "description": "Fetches specific fight object from database",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the resource",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fetched fight object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FightResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/ERR_FIGHT_NOT_FOUND"
          },
          "422": {
            "$ref": "#/components/responses/ERR_INVALID_ID"
          },
          "500": {
            "$ref": "#/components/responses/ERR_CONNECT"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "FighterResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "response": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer"
                },
                "ime": {
                  "type": "string"
                },
                "prezime": {
                  "type": "string"
                },
                "rekord": {
                  "type": "string"
                },
                "datum_rođenja": {
                  "type": "string"
                },
                "preciznost_značajnih_udaraca": {
                  "type": "string"
                },
                "broj_značajnih_udaraca_po_minuti": {
                  "type": "number"
                },
                "preciznost_rušenja": {
                  "type": "string"
                },
                "broj_rušenja_po_minuti": {
                  "type": ["number", "integer"]
                },
                "links": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "href": {
                        "type": "string"
                      },
                      "rel": {
                        "type": "string"
                      },
                      "type": {
                        "type": "string"
                      }
                    }
                  }
                },
                "prethodne_borbe": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "protivnik": {
                        "type": "string"
                      },
                      "rezultat": {
                        "type": "string"
                      },
                      "datum_borbe": {
                        "type": "string"
                      },
                      "links": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "href": {
                              "type": "string"
                            },
                            "rel": {
                              "type": "string"
                            },
                            "type": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    },
                    "required": [
                      "protivnik",
                      "rezultat",
                      "datum_borbe",
                      "links"
                    ]
                  }
                }
              },
              "required": [
                "id",
                "ime",
                "prezime",
                "rekord",
                "datum_rođenja",
                "preciznost_značajnih_udaraca",
                "broj_značajnih_udaraca_po_minuti",
                "preciznost_rušenja",
                "broj_rušenja_po_minuti",
                "links",
                "prethodne_borbe"
              ]
            }
          }
        },
        "additionalProperties": false,
        "required": ["status", "message", "response"]
      },
      "Fighter": {
        "type": "object",
        "title": "Fighter",
        "required": [
          "ime",
          "prezime",
          "rekord",
          "datum_rođenja",
          "preciznost_značajnih_udaraca",
          "broj_značajnih_udaraca_po_minuti",
          "preciznost_rušenja",
          "broj_rušenja_po_minuti",
          "prethodne_borbe"
        ],
        "properties": {
          "ime": {
            "type": "string",
            "title": "Ime",
            "description": "Ime UFC borca"
          },
          "prezime": {
            "type": "string",
            "title": "Prezime",
            "description": "Prezime UFC borca"
          },
          "rekord": {
            "type": "string",
            "pattern": "^(\\d+)-(\\d+)-(\\d+)$",
            "title": "Rekord",
            "description": "Rekord pobjeda-poraz-neriješeno u karijeri UFC borca"
          },
          "datum_rođenja": {
            "type": "string",
            "format": "date",
            "title": "Datum rođenja",
            "description": "Datum rođenja UFC borca"
          },
          "preciznost_značajnih_udaraca": {
            "type": "string",
            "title": "Preciznost značajnih udaraca",
            "description": "Postotak značajnih udaraca UFC borca u odnosu na ukupan broj udaraca"
          },
          "broj_značajnih_udaraca_po_minuti": {
            "type": "number",
            "title": "Broj značajnih udaraca po minuti",
            "description": "Ukupan broj značajnih udaraca UFC borca podijeljen s njegovim ukupnim vremenom borbe"
          },
          "preciznost_rušenja": {
            "type": "string",
            "title": "Preciznost rušenja",
            "description": "Postotak uspješnih rušenja UFC borca u odnosu na ukupan broj rušenja"
          },
          "broj_rušenja_po_minuti": {
            "type": ["number", "integer"],
            "title": "Broj rušenja po minuti",
            "description": "Ukupan broj rušenja UFC broca podijeljen s njegovim ukupnim vremenom borbe"
          },
          "prethodne_borbe": {
            "type": "array",
            "title": "Prethodne borbe",
            "description": "Podaci o prethodne dvije borbe UFC borca",
            "items": {
              "type": "object",
              "title": "Borba",
              "description": "Protivnik, rezultat i datum borbe UFC borca",
              "required": ["protivnik", "rezultat", "datum_borbe"],
              "properties": {
                "protivnik": {
                  "type": "string",
                  "title": "Ime i prezime protivnika",
                  "description": "Ime i prezime protivnika u prethodnoj borbi UFC borca"
                },
                "rezultat": {
                  "type": "string",
                  "title": "Rezultat",
                  "description": "Rezultat prethodne borbe UFC borca: pobjeda, poraz ili neriješeno"
                },
                "datum_borbe": {
                  "type": "string",
                  "format": "date",
                  "title": "Datum borbe",
                  "description": "Datum prethodne borbe UFC borca"
                }
              }
            }
          }
        }
      },
      "FighterWithId": {
        "type": "object",
        "title": "FighterWithId",
        "required": ["id"],
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "Id UFC borca"
          },
          "ime": {
            "type": "string",
            "title": "Ime",
            "description": "Ime UFC borca"
          },
          "prezime": {
            "type": "string",
            "title": "Prezime",
            "description": "Prezime UFC borca"
          },
          "rekord": {
            "type": "string",
            "pattern": "^(\\d+)-(\\d+)-(\\d+)$",
            "title": "Rekord",
            "description": "Rekord pobjeda-poraz-neriješeno u karijeri UFC borca"
          },
          "datum_rođenja": {
            "type": "string",
            "format": "date",
            "title": "Datum rođenja",
            "description": "Datum rođenja UFC borca"
          },
          "preciznost_značajnih_udaraca": {
            "type": "string",
            "title": "Preciznost značajnih udaraca",
            "description": "Postotak značajnih udaraca UFC borca u odnosu na ukupan broj udaraca"
          },
          "broj_značajnih_udaraca_po_minuti": {
            "type": "number",
            "title": "Broj značajnih udaraca po minuti",
            "description": "Ukupan broj značajnih udaraca UFC borca podijeljen s njegovim ukupnim vremenom borbe"
          },
          "preciznost_rušenja": {
            "type": "string",
            "title": "Preciznost rušenja",
            "description": "Postotak uspješnih rušenja UFC borca u odnosu na ukupan broj rušenja"
          },
          "broj_rušenja_po_minuti": {
            "type": ["number", "integer"],
            "title": "Broj rušenja po minuti",
            "description": "Ukupan broj rušenja UFC broca podijeljen s njegovim ukupnim vremenom borbe"
          },
          "prethodne_borbe": {
            "type": "array",
            "title": "Prethodne borbe",
            "description": "Podaci o prethodne dvije borbe UFC borca",
            "items": {
              "type": "object",
              "title": "Borba",
              "description": "Protivnik, rezultat i datum borbe UFC borca",
              "required": ["protivnik", "rezultat", "datum_borbe"],
              "properties": {
                "protivnik": {
                  "type": "string",
                  "title": "Ime i prezime protivnika",
                  "description": "Ime i prezime protivnika u prethodnoj borbi UFC borca"
                },
                "rezultat": {
                  "type": "string",
                  "title": "Rezultat",
                  "description": "Rezultat prethodne borbe UFC borca: pobjeda, poraz ili neriješeno"
                },
                "datum_borbe": {
                  "type": "string",
                  "format": "date",
                  "title": "Datum borbe",
                  "description": "Datum prethodne borbe UFC borca"
                }
              }
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "response": {
            "type": "null"
          }
        },
        "additionalProperties": false,
        "required": ["status", "message", "response"]
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": false,
        "required": ["status", "message"]
      },
      "FighterBriefResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "response": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer"
                },
                "ime": {
                  "type": "string"
                },
                "prezime": {
                  "type": "string"
                },
                "links": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "href": {
                        "type": "string"
                      },
                      "rel": {
                        "type": "string"
                      },
                      "type": {
                        "type": "string"
                      }
                    }
                  }
                }
              },
              "required": ["id", "ime", "prezime", "links"]
            }
          }
        },
        "additionalProperties": false,
        "required": ["status", "message", "response"]
      },
      "FightResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "response": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "borac": {
                  "type": "object",
                  "properties": {
                    "id_borca": {
                      "type": "integer"
                    },
                    "ime_borca": {
                      "type": "string"
                    },
                    "links": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "href": {
                            "type": "string"
                          },
                          "rel": {
                            "type": "string"
                          },
                          "type": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "required": ["id_borca", "ime_borca", "links"],
                  "additionalProperties": false
                },
                "protivnik": {
                  "type": "object",
                  "properties": {
                    "id_protivnika": {
                      "type": "integer"
                    },
                    "ime_protivnika": {
                      "type": "string"
                    },
                    "links": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "href": {
                            "type": "string"
                          },
                          "rel": {
                            "type": "string"
                          },
                          "type": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "required": ["id_protivnika", "ime_protivnika", "links"],
                  "additionalProperties": false
                },
                "datum_borbe": {
                  "type": "string"
                },
                "rezultat": {
                  "type": "string"
                }
              },
              "required": ["borac", "protivnik", "datum_borbe", "rezultat"]
            }
          }
        },
        "additionalProperties": false,
        "required": ["status", "message", "response"]
      },
      "PostSuccessResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "response": {
            "type": "object",
            "properties": {
              "links": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "href": {
                      "type": "string"
                    },
                    "rel": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    }
                  }
                }
              }
            },
            "required": ["links"],
            "additionalProperties": false
          }
        },
        "additionalProperties": false,
        "required": ["status", "message", "response"]
      }
    },
    "responses": {
      "ERR_CONNECT": {
        "description": "Error connecting to database",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "ERR_ADD": {
        "description": "Error while adding fighter to database!",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "ERR_OPP": {
        "description": "Opponent does not exist in database!",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "ERR_UPD": {
        "description": "Error while updating values in database!",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "ERR_FIGHTER_NOT_FOUND": {
        "description": "Fighter does not exist!",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "ERR_FIGHT_NOT_FOUND": {
        "description": "Fight does not exist!",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "ERR_DEL": {
        "description": "Error while deleting figter from database!",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "ERR_INVALID_ID": {
        "description": "Invalid fighter id!",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "ERR_NOT_IMPLEMENTED": {
        "description": "Method not implemented for requested resource",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
    }
  }
}